+++ data-cfenv-processor/src/test/java/com/ibm/cfenv/spring/boot/data/CloudantCfEnvProcessorApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ibm.cfenv.spring.boot;
+package com.ibm.cfenv.spring.boot.data;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
